refactor: move and deduplicate testutils, update imports
