+++ pkg/internal/testabilities/testservices/fixture_arc.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/defs"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/testabilities/testutils"
 	"github.com/bsv-blockchain/go-sdk/chainhash"
 	sdk "github.com/bsv-blockchain/go-sdk/transaction"
 	"github.com/go-resty/resty/v2"
@@ -222,7 +223,7 @@ func (a *arcQueryFixture) WillReturnTransactionWithMerklePath(path sdk.MerklePat
 }
 
 func (a *arcQueryFixture) WillReturnWithMindedTx() ARCQueryFixture {
-	merklePath := mockValidMerklePath(a.TB, a.txID)
+	merklePath := testutils.MockValidMerklePath(a.TB, a.txID)
 	return a.WillReturnTransactionWithMerklePath(merklePath)
 }
 
@@ -316,32 +317,5 @@ func errorResponseForStatusWithExtraInfo(httpStatus int, extraInfo string) (int,
 		"title":     title,
 		"txid":      nil,
 		"type":      "https://bitcoin-sv.github.io/arc/#/errors?id=_" + to.StringFromInteger(httpStatus),
-	}
-}
-
-func mockValidMerklePath(t testing.TB, txID string) sdk.MerklePath {
-	t.Helper()
-
-	hash, err := chainhash.NewHashFromHex(txID)
-	require.NoError(t, err)
-
-	someSecondHash, errHash := chainhash.NewHashFromHex("27a53423aa3e5d5c46bf30be53a9998dd247daf758847f244f82d430be71de6e")
-	require.NoError(t, errHash)
-
-	return sdk.MerklePath{
-		BlockHeight: 2000,
-		Path: [][]*sdk.PathElement{
-			{
-				{
-					Offset: 0,
-					Hash:   hash,
-					Txid:   to.Ptr(true),
-				},
-				{
-					Offset: 1,
-					Hash:   someSecondHash,
-				},
-			},
-		},
 	}
 }
