+++ pkg/internal/testabilities/testutils/predicates.go
